Reformated the match handler and the class helping it, tried to do it cleaner
